requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-pptx==0.6.21
python-dotenv==1.0.0 